+++ src/refreshers.py
@@ -1,12 +1,12 @@
 from datetime import datetime
+from google.appengine.api import taskqueue
 from main import app, db
 from models import Author
 from models import Publication
-import random
 from sqlalchemy import or_, func
-from google.appengine.api import taskqueue
+from sqlalchemy.orm import load_only
 
-p = 0.05
+P = 0.05
 
 @app.route('/author/refresh', methods=['GET'])
 def refresh_authors():
@@ -14,14 +14,16 @@ def refresh_authors():
   Schedules some of the out-dated authors to be refreshed.
   """
 
-  authors = Author.query.filter(Author.scholar_id != None, or_(Author.retrieved_at == None, func.now() > func.adddate(Author.retrieved_at, 30))).all()
+  OUT_DATE_LIMIT = 30
+  FETCH_LIMIT = 10000
+
+  authors = Author.query.filter(Author.scholar_id != None, func.rand() < P, or_(Author.retrieved_at == None, func.now() > func.adddate(Author.retrieved_at, OUT_DATE_LIMIT))).options(load_only('scholar_id')).limit(FETCH_LIMIT).all()
   count = 0
   for author in authors:
-    if random.random() < p:
-        queue = taskqueue.Queue('author-fetchers')
-        task = taskqueue.Task(url='/author/crawl', params={'scholar_id': author.scholar_id})
-        queue.add(task)
-        count += 1
+      queue = taskqueue.Queue('author-fetchers')
+      task = taskqueue.Task(url='/author/crawl', params={'scholar_id': author.scholar_id})
+      queue.add(task)
+      count += 1
 
   print 'Had ' + str(count) + ' authors refreshed.'
   return 'Refreshed.'
@@ -32,14 +34,16 @@ def refresh_publications():
   Schedules some of the out-dated publications to be refreshed.
   """
 
-  publications = Publication.query.filter(Publication.scholar_id != None, or_(Publication.retrieved_at == None, func.now() > func.adddate(Publication.retrieved_at, 180))).all()
+  OUT_DATE_LIMIT = 180
+  FETCH_LIMIT = 1000
+
+  publications = Publication.query.filter(Publication.scholar_id != None, func.rand() < P, or_(Publication.retrieved_at == None, func.now() > func.adddate(Publication.retrieved_at, OUT_DATE_LIMIT))).options(load_only('scholar_id')).limit(FETCH_LIMIT).all()
   count = 0
   for publication in publications:
-    if random.random() < p:
-        queue = taskqueue.Queue('publication-fetchers')
-        task = taskqueue.Task(url='/publication/crawl', params={'scholar_id': publication.scholar_id})
-        queue.add(task)
-        count += 1
+      queue = taskqueue.Queue('publication-fetchers')
+      task = taskqueue.Task(url='/publication/crawl', params={'scholar_id': publication.scholar_id})
+      queue.add(task)
+      count += 1
 
   print 'Had ' + str(count) + ' publications refreshed.'
   return 'Refreshed.'